[LF-49/batch-cron-job] batch 어플리케이션 quartz scheduler 제거
