+++ LiveFeedBatch/build.gradle
@@ -2,11 +2,10 @@ plugins {
     id 'java-library'
 }
 
-version = '0.2.3'
+version = '0.2.4'
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-batch'
-    implementation 'org.springframework.boot:spring-boot-starter-quartz'
     implementation 'org.seleniumhq.selenium:selenium-java:4.17.0'
     implementation 'mysql:mysql-connector-java:8.0.33'
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
