+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/configuration/NaverNewsCrawlJobConfiguration.java
@@ -3,7 +3,7 @@ package com.livefeed.livefeedbatch.batch.configuration;
 import com.livefeed.livefeedbatch.batch.common.dto.ItemDto;
 import com.livefeed.livefeedbatch.batch.processor.NaverNewsContentProcessor;
 import com.livefeed.livefeedbatch.batch.reader.NaverNewsUrlReader;
-import com.livefeed.livefeedbatch.batch.writer.NewsUrlWriter;
+import com.livefeed.livefeedbatch.batch.writer.NewsWriter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
@@ -16,26 +16,26 @@ import org.springframework.transaction.PlatformTransactionManager;
 
 @Configuration
 @RequiredArgsConstructor
-public class NaverNewsUrlCrawlJobConfiguration {
+public class NaverNewsCrawlJobConfiguration {
     private static final int chunkSize = 20;
 
     private final NaverNewsUrlReader naverNewsUrlReader;
     private final NaverNewsContentProcessor naverNewsContentProcessor;
-    private final NewsUrlWriter newsUrlWriter;
+    private final NewsWriter newsWriter;
 
     @Bean
-    public Job naverNewsCrawlJob(JobRepository jobRepository, PlatformTransactionManager transactionManager) throws Exception {
+    public Job naverNewsCrawlJob(JobRepository jobRepository, PlatformTransactionManager transactionManager) {
         return new JobBuilder("naverNewsCrawlJob", jobRepository)
                 .start(naverNewsCrawlStep(jobRepository, transactionManager))
                 .build();
     }
 
-    public Step naverNewsCrawlStep(JobRepository jobRepository, PlatformTransactionManager transactionManager) throws Exception {
+    public Step naverNewsCrawlStep(JobRepository jobRepository, PlatformTransactionManager transactionManager) {
         return new StepBuilder("naverNewsUrlCrawlStep", jobRepository)
                 .<String, ItemDto>chunk(chunkSize, transactionManager)
                 .reader(naverNewsUrlReader)
                 .processor(naverNewsContentProcessor)
-                .writer(newsUrlWriter)
+                .writer(newsWriter)
                 .build();
     }
 }
