+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/reader/NaverNewsUrlReader.java
@@ -5,6 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
 import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.batch.item.data.AbstractPaginatedDataItemReader;
 import org.springframework.beans.factory.annotation.Value;
@@ -108,6 +110,8 @@ public class NaverNewsUrlReader extends AbstractPaginatedDataItemReader<String> 
     }
 
     private void readArticles(WebDriver driver, List<String> articleUrls) {
+        WebDriverWait webDriverWait = new WebDriverWait(driver, Duration.ofSeconds(5));
+        webDriverWait.until(ExpectedConditions.presenceOfElementLocated(By.cssSelector(".news_list .text")));
         List<WebElement> articleList = driver.findElements(By.cssSelector(".news_list .text"));
 
         for (WebElement articleElement : articleList) {
